[FIX] make auth pakage
