+++ src/main/java/com/hongik/pcrc/allinone/auth/ui/requestBody/EmailVerifyRequest.java
@@ -1,0 +1,17 @@
+package com.hongik.pcrc.allinone.auth.ui.requestBody;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+public class EmailVerifyRequest {
+
+    //Email Info
+    private String id;
+    private String code;
+}
