--- src/main/java/com/hongik/pcrc/allinone/ui/view/ApiResponseView.java
@@ -1,17 +1,0 @@
-package com.hongik.pcrc.allinone.ui.view;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class ApiResponseView<T> {
-
-    private final T data;
-
-    public ApiResponseView(T data) {
-        this.data = data;
-    }
-}
